final back end fix up
