+++ Backend/src/controllers/user/customer_controller.js
@@ -1,11 +1,11 @@
 const User = require("../../models/customermodel");
-const AppError = require("../../utils/appError");
-const catchAsync = require("../../utils/catchAsyncError");
-const { signJwtToken } = require("../../utils/processJWT");
+const AppError = require("../../utils/Error");
+const catchAsync = require("../../utils/catcherror");
+const { signJwtToken } = require("../../utils/pJWT");
 const {
   sendResponse,
   sendResponseWithToken,
-} = require("../../utils/successResponse");
+} = require("../../utils/sucResponse");
 const constroller = require("../ge_customer");
 
 const filterBody = (reqBody, ...allowedFields) => {
