+++ Backend/src/routes/authenticationroute.js
@@ -1,12 +1,12 @@
 const express = require('express');
-const { login, register } = require('../controllers/auth/authenticatecont');
 const { userSigninInputRule, userSignupInputRule, validateInput } = require('../utils/validation');
+const { login, register } = require('../controllers/auth/authenticatecont');
+
 
 const authRouter = express.Router();
 
-authRouter.post('/login', userSigninInputRule(), validateInput, login);
-
 authRouter.post('/register', userSignupInputRule(), validateInput, register)
+authRouter.post('/login', userSigninInputRule(), validateInput, login);
 
 
 module.exports = authRouter;
