+++ Backend/src/utils/Error.js
@@ -1,0 +1,16 @@
+
+//author: mikias hailu 
+class AppError extends Error {
+
+    constructor(message, statusCode) {
+        super(message);
+
+        this.statusCode = statusCode;
+        this.isOperational = true;
+        this.status = `${statusCode}`.startsWith('4') ? 'fail' : 'error';
+       
+        Error.captureStackTrace(this, this.constructor);
+    }
+}
+
+module.exports = AppError;
