+++ Backend/src/utils/catcherror.js
@@ -1,3 +1,4 @@
+//author: mikias hailu and yared tsgie
 const catchAsync = func => {
     return (req, res, next) => {
         func(req, res, next).catch(err => next(err));
