+++ Backend/src/utils/checkRole.js
@@ -1,5 +1,6 @@
+//author: mikiashailu and yared tsgie
 const adminList = process.env.ADMIN_EMAIL ? process.env.ADMIN_EMAIL.split(',') : [];
-
+//admin
 exports.isAdmin = (user) => {
   if (adminList.includes(user.email.trim())) {
     return true;
