+++ Backend/src/utils/env.js
@@ -1,1 +1,2 @@
+//env
 exports.isDev = () => process.env.NODE_ENV == "development" ? true : false
